[State Sync] Rename "synchronizer_environment" to "test_harness".
